% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_clear_data.r
\name{clear_infomove_data}
\alias{clear_infomove_data}
\title{Clear simulation data.}
\usage{
clear_infomove_data(ssh_con = "some server", password = "some password")
}
\arguments{
\item{ssh_con}{Peregrine account to which to connect.}

\item{password}{Peregrine password.}
}
\value{
Checks out the infomove branch.
}
\description{
Clear simulation data.
}
