% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_fl_coefs.r
\name{get_fl_coefs}
\alias{get_fl_coefs}
\title{Collect data for the fitness landscape.}
\usage{
get_fl_coefs(data_path = "data", type = "info")
}
\arguments{
\item{data_path}{Where to search for data. Must be run in the infomove folder on the cluster.}

\item{type}{The type of simulation to search for.}
}
\value{
Prints the coefficients of a and b to file.
}
\description{
Collect data for the fitness landscape.
}
