% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_summary.r
\name{get_summary}
\alias{get_summary}
\title{Get summary data from the infomove simulation.}
\usage{
get_summary(
  ssh_con = "some server",
  password = "some password",
  type = "noinfo"
)
}
\arguments{
\item{ssh_con}{The ssh address. Typically a p-number at peregrine.hpc.rug.nl. This function is meant to be used within infomove.}

\item{password}{Connection password.}

\item{type}{The simulation type. Currently "info" or "noinfo". May also be all.}
}
\value{
Nothing. Gets summary data from the infomove simulation.
}
\description{
Get summary data from the infomove simulation.
}
