% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make_land.r
\name{make_land}
\alias{make_land}
\title{Make landscape values.}
\usage{
make_land(size = 5000, phi_peaks = 100, rho_error = 0.1)
}
\arguments{
\item{size}{Total number of landscape patches.}

\item{phi_peaks}{Number of resource peaks.}

\item{rho_error}{Error around the sine function.}
}
\value{
A vector of landscape values.
}
\description{
Make landscape values.
}
