% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_fitness_landscape.r
\name{print_fitness_landscape}
\alias{print_fitness_landscape}
\title{Collect data for the fitness landscape.}
\usage{
print_fitness_landscape(data_path = "data", type = "info")
}
\arguments{
\item{data_path}{Where to search for data. Must be run in the infomove folder on the cluster.}

\item{type}{The type of simulation to search for.}
}
\value{
Prints figures and saves data providing a global summary of all data in the data folder.
}
\description{
Collect data for the fitness landscape.
}
