% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_global_summary.r
\name{print_global_summary}
\alias{print_global_summary}
\title{Print a global summary.}
\usage{
print_global_summary(data_path = "data", type = "info")
}
\arguments{
\item{data_path}{Where to search for data.}

\item{type}{The type of simulation to search for.}
}
\value{
Prints figures and saves data providing a global summary of all data in the data folder.
}
\description{
Print a global summary.
}
