% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_run_infomove.r
\name{run_infomove}
\alias{run_infomove}
\title{Run the infomove simulation.}
\usage{
run_infomove(
  type = "noinfo",
  phi = 15,
  rho = 0.09,
  gens = "100000",
  timesteps = 100,
  init_d = 0.99,
  replicates = 10
)
}
\arguments{
\item{type}{The simulation type. Currently "info" or "noinfo".}

\item{phi}{Value (vector) of PHI values.}

\item{rho}{Value (vector) of RHO values.}

\item{gens}{Maximum number of generations.}

\item{timesteps}{Timesteps per generation.}

\item{init_d}{Initial value of giving up density D.}

\item{replicates}{How many replicates to run.}
}
\value{
Nothing. Runs the infomove simulation.
}
\description{
Run the infomove simulation.
}
