% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_run_fitness_landscape.r
\name{run_infomove_fl}
\alias{run_infomove_fl}
\title{Get the local fitness landscape.}
\usage{
run_infomove_fl(
  type = "noinfo",
  phi = 15,
  rho = 0.1,
  timesteps = 500,
  a_res = 1,
  b_res = 1,
  M_res = 2,
  leader_choices = c(1, 2, 5),
  gradient = 0.5,
  gradient_m = 2,
  n_rep = 10
)
}
\arguments{
\item{type}{The simulation type. Currently "info" or "noinfo".}

\item{phi}{Value (vector) of PHI values.}

\item{rho}{Value (vector) of RHO values.}

\item{timesteps}{Timesteps per generation.}

\item{a_res}{The resident a strategy.}

\item{b_res}{The resident b strategy.}

\item{M_res}{The resident M strategy.}

\item{leader_choices}{How many leaders can be assessed.}

\item{gradient}{The gradient increment by which to shift the parameters.}

\item{gradient_m}{The gradient increment for M}

\item{n_rep}{How many replicates to run.}
}
\value{
Nothing. Runs the infomove simulation for local fitness landscape.
}
\description{
Get the local fitness landscape.
}
