% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_select_version.r
\name{select_which_infomove}
\alias{select_which_infomove}
\title{Select infomove version.}
\usage{
select_which_infomove(
  ssh_con = "some server",
  password = "some password",
  branch = "allow_D"
)
}
\arguments{
\item{ssh_con}{Peregrine account to which to connect.}

\item{password}{Peregrine password.}

\item{branch}{Branch to be checked out.}
}
\value{
Checks out the infomove branch.
}
\description{
Select infomove version.
}
