% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_run_infomove.r
\name{run_infomove}
\alias{run_infomove}
\title{Run the infomove simulation.}
\usage{
run_infomove(
  ssh_con = "some server",
  password = "some password",
  type = "noinfo",
  phi = 15,
  rho = 0.1,
  gens = "100000",
  timesteps = 100,
  init_d = 0.5,
  leader_choices = 2,
  rep_n = 5
)
}
\arguments{
\item{ssh_con}{The ssh address. Typically a p-number at peregrine.hpc.rug.nl}

\item{password}{Connection password.}

\item{type}{The simulation type. Currently "info" or "noinfo".}

\item{phi}{Value (vector) of PHI values.}

\item{rho}{Value (vector) of RHO values.}

\item{gens}{Maximum number of generations.}

\item{timesteps}{Timesteps per generation.}

\item{init_d}{Initial Value of aspiration level D.}

\item{leader_choices}{How many leaders can be assessed.}

\item{rep_n}{How many rep_n to run.}
}
\value{
Nothing. Runs the infomove simulation.
}
\description{
Run the infomove simulation.
}
